[#13]refactor: verion controller 파일 네이밍을 camelCase로 변경
